+++ google_play_scraper/features/reviews.py
@@ -1,4 +1,5 @@
 import json
+from typing import Optional
 
 from google_play_scraper import Sort
 from google_play_scraper.constants.element import ElementSpecs
@@ -7,10 +8,16 @@ from google_play_scraper.constants.url import Formats
 from google_play_scraper.utils.request import post
 
 
-def _fetch_review_items(url, app_id, sort, count, pagination_token):
+def _fetch_review_items(url, app_id, sort, count, filter_score_with, pagination_token):
     dom = post(
         url,
-        Formats.ReviewPayload.build(app_id, sort, count, pagination_token),
+        Formats.ReviewPayload.build(
+            app_id,
+            sort,
+            count,
+            "null" if filter_score_with is None else filter_score_with,
+            pagination_token,
+        ),
         {"content-type": "application/x-www-form-urlencoded"},
     )
 
@@ -19,9 +26,10 @@ def _fetch_review_items(url, app_id, sort, count, pagination_token):
     return json.loads(match[0][2])[0], json.loads(match[0][2])[-1][-1]
 
 
-def reviews(app_id, lang="en", country="us", sort=Sort.NEWEST, count=100):
-    # type: (str, str, str, Sort, int) -> list
-    # TODO filtering with rating
+def reviews(
+    app_id, lang="en", country="us", sort=Sort.NEWEST, count=100, filter_score_with=None
+):
+    # type: (str, str, str, Sort, int, Optional[int]) -> list
     # TODO filtering with device model
     # TODO reply data
     # TODO refactoring
@@ -39,7 +47,7 @@ def reviews(app_id, lang="en", country="us", sort=Sort.NEWEST, count=100):
 
     while True:
         review_items, pagination_token = _fetch_review_items(
-            url, app_id, sort, _count, pagination_token
+            url, app_id, sort, _count, filter_score_with, pagination_token
         )
 
         for review in review_items: